build: use each external library by makeavailable
This is to avoid having to know they underlying structure of library
codebase is depending on, and to let the library add and install target.

In the previous configuration requirements were specific and everything
was header-only used, however as there can not be more than one install
TARGETS cmake command in a cmake file, we have switched to using
FetchContent_MakeAvailable command (instead of manually populating and
etc).